fix: logout function did not redirect
Fix a bug where the provided logout function did not redirect the
browser to the authentication providers logout endpoint.
